agents/nautilus/pyproject.toml
[project]
name = "nautilus-agent"
version = "0.1.0"
description = "NautilusTrader IB adapter agent — replaces C++ TWSClient as the primary IBKR connection"
requires-python = ">=3.12"
dependencies = [
    "nautilus_trader[ib]==1.224.0",
    "nats-py>=2.7.0",
    "google-protobuf>=4.25",
    "grpcio-tools>=1.62",
    "structlog>=24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "mypy>=1.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nautilus_agent"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "mypy>=1.9",
]
